Refactor tests and add domain validation
Two main changes have been made in this commit. The first one is a refactor in the 'CategoryTest' class, replacing hardcoded inline data with a generator function for a specific range of invalid names. This change optimizes the code by reducing redundancy and improving flexibility for future test adjustments. In addition, a new domain validation functionality has been added, which allows for universal non-null validation on any field. This significantly enhances the code's robustness as it ensures that important fields in the Domain Entities are not left null.